+++ epd-ship/src/main/java/dk/dma/epd/ship/gui/route/strategic/SendStrategicRouteDialog.java
@@ -99,7 +99,7 @@ public class SendStrategicRouteDialog extends ComponentDialog implements ActionL
     private boolean loading;
     private boolean wasVisible;
 
-    private ArrayList<Integer> mmsiList = new ArrayList<Integer>();
+    private ArrayList<Long> mmsiList = new ArrayList<Long>();
 
 
     IdentityViewer viewer;
